+++ gTeamProject/__init__.py
@@ -2,4 +2,4 @@ from __future__ import absolute_import
 from .celery import app as celery_app
 
 
-__all__ = ['celery_app']
+__all__ = ["celery_app"]
